Create Argument.cpp
This Program shows how we can create  Argument with the help of Structures in C++ or C (both languages is capable of this programming). As, You can see i have used structures and a function  name void which does not return value and a reference value(&) to avoid copying in C++ or C  Programming. 
NOTE: Please Cross check that how you are declaring the values and  as it is you have to write the Print format and the values in main function in series.